fix: sort/mail button
