+++ src/mail.js
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
 USER_ID: `QWOEOQWE`, //userID
 TEMPLATE_ID: `QPWEOPQWE`,         //templateID
 }
